requirements.txt
streamlit>=1.31.0
google-adk>=0.1.0
python-dotenv>=1.0.0
